Fixed so that config file is in yaml format and can easily be read in the correct format to read_file program.
