--- src/config_test.txt
@@ -1,21 +1,0 @@
-copper.vasp copper2.vasp
-5 10 15 20
-10 20 30 40
-5 10 15 20
-False
-0 0.5;0 0.5;0 0
-0 50 300 100
-100
-5
-1000
-
-aluminum.vasp
-2 4 8 16
-10 15 25 50
-5 10 15 20
-False
-0 0.5;0 0.5;0 0
-0 50 300 20
-100
-5
-1000
