+++ src/main.py
@@ -1,5 +1,5 @@
 from build import *
-from read_file import *
+from read_file_old import *
 from md import * 
 
 def main():
