--- src/read_file.log
@@ -1,6 +1,0 @@
-2024-10-14 10:38:37,739 - __main__ - INFO - Reading config file...
-2024-10-14 10:38:37,740 - __main__ - INFO - Converting data for config block 1...
-2024-10-14 10:38:37,740 - __main__ - INFO - Succeded converting data for block 1!
-2024-10-14 10:38:37,740 - __main__ - INFO - Converting data for config block 2...
-2024-10-14 10:38:37,740 - __main__ - INFO - Succeded converting data for block 2!
-2024-10-14 10:38:37,741 - __main__ - INFO - Done!
